Added three separate components for audio, video and selfie
