+++ react/features/base/jwt/actions.js
@@ -15,8 +15,10 @@ export function setJWT(jwt: ?string) {
     return {
         type: SET_JWT,
         //video
-        // jwt: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdWQiOiJqaXRzaSIsImlzcyI6InBvbHl0b2siLCJzdWIiOiJwb2x5dG9rIiwicm9vbSI6IioiLCJzZWxmaWUiOiJWIiwibmJmIjoxNjYyMDk1MTQ2LCJleHAiOjE2NjIwOTg4NjZ9.ZmoPW-s44QaFV7LfezThl9NClDY3xFEf5IMlBb0nUjw'
+        jwt: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdWQiOiJqaXRzaSIsImlzcyI6InBvbHl0b2siLCJzdWIiOiJwb2x5dG9rIiwicm9vbSI6IioiLCJzZWxmaWUiOiJWIiwibmJmIjoxNjYyMDk1MTQ2LCJleHAiOjE2NjIwOTg4NjZ9.ZmoPW-s44QaFV7LfezThl9NClDY3xFEf5IMlBb0nUjw'
         //audio
-        jwt: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdWQiOiJqaXRzaSIsImlzcyI6InBvbHl0b2siLCJzdWIiOiJwb2x5dG9rIiwicm9vbSI6IioiLCJzZWxmaWUiOiJBIiwibmJmIjoxNjYyMDk1MTQ2LCJleHAiOjE2NjIwOTg4NjZ9.pFpcOG0DiSckpve-DSnrn9-PMld5iaTXATWllJoqESc'
+        // jwt: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdWQiOiJqaXRzaSIsImlzcyI6InBvbHl0b2siLCJzdWIiOiJwb2x5dG9rIiwicm9vbSI6IioiLCJzZWxmaWUiOiJBIiwibmJmIjoxNjYyMDk1MTQ2LCJleHAiOjE2NjIwOTg4NjZ9.pFpcOG0DiSckpve-DSnrn9-PMld5iaTXATWllJoqESc'
+        //picture
+        // jwt: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdWQiOiJqaXRzaSIsImlzcyI6InBvbHl0b2siLCJzdWIiOiJwb2x5dG9rIiwicm9vbSI6IioiLCJzZWxmaWUiOiJQIiwibmJmIjoxNjYyMDk1MTQ2LCJleHAiOjE2NjIwOTg4NjZ9.RBPk5Oi0t3MK1kzQtI03WPX8YnmoQjYjgDI9kfiqxFw'
     };
 }
